Fixed TransazioniMongoService

+++ applicazione-bancaria/src/main/java/eu/tasgroup/applicativo/service/TransazioniMongoService.java
@@ -12,7 +12,7 @@ public interface TransazioniMongoService {
 	List<TransazioniMongo> findAll();
 	int numeroTransazioniPerTipo(TipoTransazione tipo);
 	double calcoloMediaTransazioniPerCliente(int codCliente);
-	int numeroMedioTransazioniPerCliente();
+	double numeroMedioTransazioniPerCliente();
 	double totaleImportoPerMese(int mese);  
 	void deleteTransazioneMongo(TransazioniMongo transazione);
